move weighted label to its own file
