+++ dlib/cuda/cuda_dlib.h
@@ -6,6 +6,7 @@
 
 #include "tensor.h"
 #include "../geometry/rectangle.h"
+#include "../dnn/misc.h"
 
 namespace dlib
 {
